updated script to have checks before removing files; it also installs curl on debian systems if its not already instaleld

+++ scripts/get_nvim.sh
@@ -1,10 +1,24 @@
 #!/bin/sh
 
-# remove previous install if it exists
-sudo rm -rf /opt/nvim
+# installs curl if it is not already installed
+if ! [ -x curl ]
+then
+    sudo apt-get install curl
+fi
+
+# remove previous nvim install if it exists
+if [ -d /opt/nvim ]
+then
+    echo "Removing previous nvim installation: /opt/nvim"
+    sudo rm -rf /opt/nvim
+fi
 
 # remove previous install symbolic link if one exists
-rm /usr/local/bin/nvim
+if [ -e /usr/local/bin/nvim ]
+then
+    echo "Removing symbolic link: /usr/local/bin/nvim"
+    rm /usr/local/bin/nvim
+fi
 
 # downloads and extracts nvim to /opt/nvim-linux64
 curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz
